refactor: common 엔티티 경로 이동
